% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_est.R
\docType{class}
\name{int_est-class}
\alias{initialize,}
\alias{int_est-class}
\alias{int_est-method}
\title{An class called int_est, with subclasses trapezoid and simpson}
\description{
Object of class \code{integrate_it} take information about a function to perform integration estimation functions
}
\details{
An item of class `int_est' has the following slots:
\itemize
\item \code{x_values} a vector containing the x values
\item \code{y_values} a vector containing the evaluated values (f(x) = y)
\item \code{estimation} a numeric object contain the integration estimate
}
\author{
Scott Solomon \email{scott.solomon@wustl.edu}
}

