\name{integrate_it-package}
\alias{integrate_it-package}
\alias{integrate_it}
\docType{package}
\title{
\packageTitle{integrate_it}
}
\description{
\packageDescription{integrate_it}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{integrate_it}
\packageIndices{integrate_it}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{integrate_it}

Maintainer: \packageMaintainer{integrate_it}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
