% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,int_est.Trapezoid-method}
\alias{plot}
\alias{plot,int_est.Trapezoid-method}
\title{plots the estimatations for an object of class int_est}
\usage{
\S4method{plot}{int_est.Trapezoid}(x = NULL, y = x, ...)
}
\arguments{
\item{int_est_object}{An object of class int_est}
}
\value{
plots the estimation trapezoids and parabolas for objects of class int_est
 \item{plot}{the plot of the estimations}
}
\description{
plots the estimation trapezoids and parabolas for objects of class int_est
}
\note{
This function gives a visual display of the simpson and trapezoid rules
}
\author{
Scott Solomon
}

