% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,int_est-method}
\alias{print}
\alias{print,int_est-method}
\title{prints the estimatation for an object of class int_est}
\usage{
\S4method{print}{int_est}(x)
}
\arguments{
\item{int_est_object}{An object of class int_est}
}
\value{
Prints the estimated value for the int_est object
 \item{estimation}{the inputed x values}
}
\description{
prints the estimatation for an object of class int_est.Trapezoid and int_est.Simpson
}
\note{
This is a simple print function
}
\examples{

test_int_est<-new("int_est.Trapezoid, x_values = c(0:10), y_values = c(0:10), estimate = 50)
print(test_int_est)
test_int_est2<-new("int_est.Simpson, x_values = c(0:10), y_values = c(0:10), estimate = 50)
print(test_int_est)
}
\author{
Scott Solomon
}

